+++ snulbug/__init__.py
@@ -10,7 +10,14 @@ from .learn import amend_mcp_policy, learn_mcp_policy
 from .middleware import LuaConfig, LuaMiddleware
 from .presets import McpPolicyOptions, copy_builtin_preset, generate_mcp_preset, list_builtin_presets
 from .promotion import compare_decisions, diff_policies
-from .proxy import FacadeUpstream, McpFacadeProxyApp, ReverseProxyApp, create_proxy_application, run_proxy
+from .proxy import (
+    FacadeUpstream,
+    ManagedStdioMcpClient,
+    McpFacadeProxyApp,
+    ReverseProxyApp,
+    create_proxy_application,
+    run_proxy,
+)
 from .quickstart import create_mcp_quickstart
 from .recorder import append_record, load_record_log, record_audit_event, record_policy_request, replay_record_log
 from .redaction import RedactionConfig, append_audit_event, build_audit_event, redact_secrets
@@ -35,6 +42,7 @@ __all__ = [
     "MemoryStateStore",
     "McpPolicyOptions",
     "FacadeUpstream",
+    "ManagedStdioMcpClient",
     "McpFacadeProxyApp",
     "RedisStateStore",
     "RedactionConfig",